refactor(ddgi): split DDGI shader interop into data header + HF helpers
Move the DDGI constant-buffer struct out of
ShaderInterop_Renderer.h into its own file,
matching how ShaderTerrain / ShaderVoxelGrid are
organized, and follow the engine's
ShaderInterop_*.h (shared data) + *HF.hlsli (GPU
helper functions) convention.

- ShaderInterop_DDGI.h becomes a leaf data header:
  constants, DDGI_CASCADE_COUNT, ShaderDDGI (+
  Cascade), push constants and packed ray/variance
  records. It no longer includes
  ShaderInterop_Renderer.h or references
  GetScene().
- ShaderInterop_Renderer.h includes the data
  header and holds `ShaderDDGI ddgi;` instead of a
  large inline struct.
- New ddgiHF.hlsli holds every GetScene()-based
  helper (probe addressing,
  ddgi_sample_irradiance, MultiscaleMeanEstimator,
  depth-border table).
- DDGIProbe and its flags stay in
  ShaderInterop_Renderer.h: they depend on the SH
  namespace defined there, so moving them would
  create a circular include.
- Shaders that call the helpers include
  ddgiHF.hlsli; objectHF and
  ddgi_indirectprepareCS use only the data structs
  and keep ShaderInterop_DDGI.h.
