+++ WickedEngine/shaders/rtreflectionCS.hlsl
@@ -9,7 +9,7 @@
 #include "raytracingHF.hlsli"
 #include "stochasticSSRHF.hlsli"
 #include "lightingHF.hlsli"
-#include "ShaderInterop_DDGI.h"
+#include "ddgiHF.hlsli"
 
 PUSHCONSTANT(postprocess, PostProcess);
